Adds try/catch locals rewriter and refactors implicit method block try/catch logic
